src/requirements.txt
Django==3.1.2
psycopg2-binary==2.8.6
django-allauth==0.43.0
